Updated sd card reader spi device to handle 32 bit addresses.

+++ src/kernel/spidevice.c
@@ -19,7 +19,7 @@ typedef struct {
 	KStr mountPoint;
 	SdCard sdCard; // type set to SdTypeBadCard when none mounted
 	uint8_t cache[SdBlockSize];
-	uint16_t cacheBlock; // for now set to UINT16_MAX to indicate empty cache
+	uint32_t cacheBlock; // for now set to UINT32_MAX to indicate empty cache
 } SpiDeviceSdCardReaderData;
 
 typedef struct {
@@ -160,7 +160,7 @@ bool spiDeviceSdCardReaderMount(SpiDeviceId id, const char *mountPoint) {
 	}
 
 	// Mark cache block as undefined (before we even register read/write functors to be safe).
-	spiDevices[id].d.sdCardReader.cacheBlock=UINT16_MAX;
+	spiDevices[id].d.sdCardReader.cacheBlock=UINT32_MAX;
 
 	// Add block device at given point mount
 	KernelFsFileOffset size=4096; // TODO: this properly - we have notes on how to find this from the card itself
@@ -220,25 +220,24 @@ KernelFsFileOffset spiDeviceSdCardReaderReadFunctor(KernelFsFileOffset addr, uin
 	if (id>=SpiDeviceIdMax || spiDeviceGetType(id)!=SpiDeviceTypeSdCardReader || spiDevices[id].d.sdCardReader.sdCard.type==SdTypeBadCard)
 		return 0; // TODO: or -1 for error?
 
-	// Note: special case for initial block==UINT16_MAX as we use this as a special 'no cache' flag, so always have to re-read.
-	// (we can at least avoid re-reading within this function, just not between calls)
-	// TODO: fix the inefficiency noted above
-	uint16_t initialBlock=addr/SdBlockSize;
-	if (initialBlock==UINT16_MAX && spiDevices[id].d.sdCardReader.cacheBlock==UINT16_MAX)
+	// Note: special case for initial block==UINT32_MAX as we use this as a special 'no cache' flag, so always have to re-read.
+	// In reality, as block size is fixed at 2^9=512, block must be less than 2^(32-9)=2^23.
+	// This means block numbers can never be UINT32_T so this should be safe.
+	uint32_t initialBlock=addr/SdBlockSize;
+	if (initialBlock==UINT16_MAX && spiDevices[id].d.sdCardReader.cacheBlock==UINT32_MAX)
 		spiDevices[id].d.sdCardReader.cacheBlock=0; // force read in first iteration, but none thereafter
 
 	// First read up to first block boundary (or less if len is low enough).
 	KernelFsFileOffset readCount;
 	for(readCount=0; readCount<len; ++readCount,++addr) {
 		// Do we need to read a new block?
-		uint16_t block=addr/SdBlockSize;
-
+		uint32_t block=addr/SdBlockSize;
 		if (block!=spiDevices[id].d.sdCardReader.cacheBlock && !sdReadBlock(&spiDevices[id].d.sdCardReader.sdCard, block, spiDevices[id].d.sdCardReader.cache))
 			break;
 		spiDevices[id].d.sdCardReader.cacheBlock=block;
 
 		// Copy byte into users array
-		uint16_t offset=addr-block*SdBlockSize;
+		uint16_t offset=addr-block*SdBlockSize; // should be <512 so can fit in 16 bit not full 32
 		data[readCount]=spiDevices[id].d.sdCardReader.cache[offset];
 	}
